feat: disable hardware options insted of hiding the tab entirely
+ strikes a balance between visibility of options & UX
